Create wdcount.py
Use a dictionary to count the number of occurrences of each word (ignoring case) in a file. Note that the input file name is given from the command line. This program firstly prints each word which has more than 4 characters and ends with the suffix "in", along with its number of occurrences (in descending order). That is, the most frequent word should be printed first. Next, this program prints the words that don't end with the suffix "in". Lastly, this program prints the total number of words in the file.